third_party/blink/web_tests/http/tests/security/offscreen-canvas-read-blocked-by-setting.html

<!DOCTYPE html>
<title>Block reading offscreen canvas via StrictCanvasTainting setting</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitDisableReadingFromCanvas", true);

var image = new Image();
var offCanvas = new OffscreenCanvas(100, 100);
var context = offCanvas.getContext('2d');

test(function () {
    assert_throws_dom(
        "SecurityError",
        function() {
            var imageData = context.getImageData(0, 0, 100, 100);
            assert_unreached('getImageData should throw');
        });
}, 'getImageData');

test(function () {
    assert_throws_dom(
        "SecurityError",
        function() {
            context.drawImage(image, 0, 0, 100, 100);
            var imageData = context.getImageData(0, 0, 100, 100);
            assert_unreached('getImageData should throw');
        });
}, 'getImageData');

</script>
